requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
asyncio-dgram==1.2.0
attrs==20.3.0
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
discord.py==1.7.0
dnspython==1.15.0
dnspython3==1.15.0
google-api-core==1.26.3
google-api-python-client==2.14.0
google-auth==1.30.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.53.0
greenlet==1.1.0
httplib2==0.19.1
idna==2.10
IMDbPY==2021.4.18
lxml==4.6.3
mcstatus==5.1.2
multidict==5.1.0
oauth2client==4.1.3
oauthlib==3.1.0
packaging==20.9
protobuf==3.16.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyDrive==1.3.1
pyparsing==2.4.7
python-dotenv==0.17.0
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
SQLAlchemy==1.4.15
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.4
yarl==1.6.3
